3.6.13 abstract interpreter implementation
