+++ package.json
@@ -34,7 +34,7 @@
     "crlf-normalize": "^1.0.20",
     "css-color-names": "^1.0.1",
     "glob": "^11.0.1",
-    "greyhack-message-hook-client": "~0.5.0",
+    "greyhack-message-hook-client": "~0.6.0",
     "greybel-gh-mock-intrinsics": "~5.5.12",
     "greybel-intrinsics": "~5.5.5",
     "greyscript-core": "~2.5.5",
